refactor(test): change path of test:artifact away from mocha
