+++ config/jest/jest.artifact-es-bundle-core.config.js
@@ -3,5 +3,5 @@ const path = require('path');
 module.exports = {
   rootDir: path.join(__dirname, '..', '..'),
   testEnvironment: 'jsdom',
-  testMatch: ['**/test/mocha/build-artifacts/es-bundle-core.js'],
+  testMatch: ['**/test/build-artifacts/es-bundle-core.js'],
 };
