+++ test/build-artifacts/es-bundle-core.js
@@ -1,5 +1,5 @@
 import expect from "expect"
-import SwaggerUI from "../../../dist/swagger-ui-es-bundle-core"
+import SwaggerUI from "../../dist/swagger-ui-es-bundle-core"
 
 describe("webpack browser es-bundle-core build", () => {
   it("should export a function for es-bundle-core", () => {
