+++ test/build-artifacts/umd.js
@@ -1,5 +1,5 @@
 import expect from "expect"
-import SwaggerUI from "../../../dist/swagger-ui-bundle"
+import SwaggerUI from "../../dist/swagger-ui-bundle"
 
 describe("webpack browser umd build", () => {
   it("should export a function for (umd) bundle", () => {
